ImmutablePlatformModule is now an interface
